+++ mediapicker/src/main/java/mediapicker/CropListener.java
@@ -1,0 +1,12 @@
+package mediapicker;
+
+/**
+ * @author TUNGDX
+ */
+
+/**
+ * Listener used to for crop photo.
+ */
+public interface CropListener {
+  public void onSuccess(MediaItem mediaItem);
+}
