+++ mediapicker/src/main/java/mediapicker/MediaAdapter.java
@@ -1,4 +1,4 @@
-package vn.tungdx.mediapicker;
+package mediapicker;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -9,14 +9,20 @@ import android.view.ViewGroup;
 import android.view.ViewGroup.LayoutParams;
 import android.widget.AbsListView.RecyclerListener;
 import android.widget.RelativeLayout;
-
 import java.util.ArrayList;
 import java.util.List;
+import mediapicker.imageloader.MediaImageLoader;
+import mediapicker.utils.MediaUtils;
+import mediapicker.widget.PickerImageView;
+import vn.tungdx.mediapicker.R;
 
-import vn.tungdx.mediapicker.imageloader.MediaImageLoader;
-import vn.tungdx.mediapicker.utils.MediaUtils;
-import vn.tungdx.mediapicker.widget.PickerImageView;
+/**
+ * @author TUNGDX
+ */
 
+/**
+ * Adapter for display media item list.
+ */
 public class MediaAdapter extends CursorAdapter implements RecyclerListener {
   private int mMediaType;
   private MediaImageLoader mMediaImageLoader;
@@ -85,6 +91,12 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     return mMediaListSelected.size() > 0;
   }
 
+  /**
+   * Check media uri is selected or not.
+   *
+   * @param uri Uri of media item (photo, video)
+   * @return true if selected, false otherwise.
+   */
   public boolean isSelected(Uri uri) {
     if (uri == null) return false;
     for (MediaItem item : mMediaListSelected) {
@@ -93,10 +105,21 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     return false;
   }
 
+  /**
+   * Check {@link MediaItem} is selected or not.
+   *
+   * @param item {@link MediaItem} to check.
+   * @return true if selected, false otherwise.
+   */
   public boolean isSelected(MediaItem item) {
     return mMediaListSelected.contains(item);
   }
 
+  /**
+   * Set {@link MediaItem} selected.
+   *
+   * @param item {@link MediaItem} to selected.
+   */
   public void setMediaSelected(MediaItem item) {
     syncMediaSelectedAsOptions();
     if (!mMediaListSelected.contains(item)) mMediaListSelected.add(item);
@@ -110,6 +133,11 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     }
   }
 
+  /**
+   * If item selected then change to unselected and unselected to selected.
+   *
+   * @param item Item to update.
+   */
   public void updateMediaSelected(MediaItem item, PickerImageView pickerImageView) {
     if (mMediaListSelected.contains(item)) {
       mMediaListSelected.remove(item);
@@ -129,14 +157,25 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     }
   }
 
+  /**
+   * @return List of {@link MediaItem} selected.
+   */
   public List<MediaItem> getMediaSelectedList() {
     return mMediaListSelected;
   }
 
+  /**
+   * Set list of {@link MediaItem} selected.
+   */
   public void setMediaSelectedList(List<MediaItem> list) {
     mMediaListSelected = list;
   }
 
+  /**
+   * Whether clear or not media selected as options.
+   *
+   * @return true if clear, false otherwise.
+   */
   private boolean syncMediaSelectedAsOptions() {
     switch (mMediaType) {
       case MediaItem.PHOTO:
@@ -158,10 +197,14 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     return false;
   }
 
+  /**
+   * {@link MediaItem#VIDEO} or {@link MediaItem#PHOTO}
+   */
   public void setMediaType(int mediaType) {
     mMediaType = mediaType;
   }
 
+  // set numcols
   public void setNumColumns(int numColumns) {
     mNumColumns = numColumns;
   }
@@ -170,6 +213,7 @@ public class MediaAdapter extends CursorAdapter implements RecyclerListener {
     return mNumColumns;
   }
 
+  // set photo item height
   public void setItemHeight(int height) {
     if (height == mItemHeight) {
       return;
