+++ mediapicker/src/main/java/mediapicker/MediaSelectedListener.java
@@ -1,0 +1,16 @@
+package mediapicker;
+
+import java.util.List;
+
+/**
+ * @author TUNGDX
+ */
+
+/**
+ * Listener for select media item.
+ */
+public interface MediaSelectedListener {
+  public void onHasNoSelected();
+
+  public void onHasSelected(List<MediaItem> mediaSelectedList);
+}
