+++ mediapicker/src/main/java/mediapicker/activities/FragmentHost.java
@@ -1,0 +1,10 @@
+package mediapicker.activities;
+
+import mediapicker.imageloader.MediaImageLoader;
+
+/**
+ * Created by TungDX
+ */
+public interface FragmentHost {
+  MediaImageLoader getImageLoader();
+}
