+++ mediapicker/src/main/java/mediapicker/imageloader/MediaImageLoader.java
@@ -1,0 +1,11 @@
+package mediapicker.imageloader;
+
+import android.net.Uri;
+import android.widget.ImageView;
+
+/**
+ * @author TUNGDX
+ */
+public interface MediaImageLoader {
+  void displayImage(Uri uri, ImageView imageView);
+}
