--- mediapicker/src/main/java/vn/tungdx/mediapicker/CropListener.java
@@ -1,5 +1,0 @@
-package vn.tungdx.mediapicker;
-
-public interface CropListener {
-    public void onSuccess(MediaItem mediaItem);
-}
