--- mediapicker/src/main/java/vn/tungdx/mediapicker/MediaSelectedListener.java
@@ -1,9 +1,0 @@
-package vn.tungdx.mediapicker;
-
-import java.util.List;
-
-public interface MediaSelectedListener {
-    public void onHasNoSelected();
-
-    public void onHasSelected(List<MediaItem> mediaSelectedList);
-}
