--- mediapicker/src/main/java/vn/tungdx/mediapicker/activities/FragmentHost.java
@@ -1,7 +1,0 @@
-package vn.tungdx.mediapicker.activities;
-
-import vn.tungdx.mediapicker.imageloader.MediaImageLoader;
-
-public interface FragmentHost {
-    MediaImageLoader getImageLoader();
-}
