--- mediapicker/src/main/java/vn/tungdx/mediapicker/activities/MediaPickerErrorDialog.java
@@ -1,40 +1,0 @@
-package vn.tungdx.mediapicker.activities;
-
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.DialogInterface.OnClickListener;
-import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-
-import vn.tungdx.mediapicker.R;
-
-public class MediaPickerErrorDialog extends DialogFragment {
-    private String mMessage;
-    private OnClickListener mOnPositionClickListener;
-
-    public static MediaPickerErrorDialog newInstance(String msg) {
-        MediaPickerErrorDialog dialog = new MediaPickerErrorDialog();
-        Bundle bundle = new Bundle();
-        bundle.putString("msg", msg);
-        dialog.setArguments(bundle);
-        return dialog;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        mMessage = getArguments().getString("msg");
-    }
-
-    @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        return new AlertDialog.Builder(getActivity())
-                .setMessage(mMessage)
-                .setPositiveButton(R.string.ok, mOnPositionClickListener)
-                .create();
-    }
-
-    public void setOnOKClickListener(OnClickListener mOnClickListener) {
-        this.mOnPositionClickListener = mOnClickListener;
-    }
-}
