--- mediapicker/src/main/java/vn/tungdx/mediapicker/imageloader/MediaImageLoader.java
@@ -1,8 +1,0 @@
-package vn.tungdx.mediapicker.imageloader;
-
-import android.net.Uri;
-import android.widget.ImageView;
-
-public interface MediaImageLoader {
-    void displayImage(Uri uri, ImageView imageView);
-}
