+++ mediapicker/src/main/res/layout/fragment_mediapicker.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <vn.tungdx.mediapicker.widget.HeaderGridView
+    <mediapicker.widget.HeaderGridView
         android:id="@id/grid"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
